feat(control): add runtime DNS accounting and statistics tracking
- Implemented runtime DNS accounting in `runtime_dns_accounting.go` to track upload and download traffic for DNS packets.
- Introduced `runtime_stats.go` to manage runtime statistics, including upload/download rates and total traffic.
- Added tests in `runtime_stats_test.go` to ensure accurate tracking of traffic and proper functionality of the new features.
- Created mock response writers and connections for testing DNS traffic handling.
- Ensured that traffic recording is isolated per instance of `ControlPlane`.

+++ control/dns_controller_reload_test.go
@@ -178,7 +178,7 @@ func TestDnsController_ReuseForReloadUpdatesBehaviorConfig(t *testing.T) {
 	require.NoError(t, err)
 	require.NotNil(t, reused)
 
-	require.Equal(t, uint16(dnsmessage.TypeAAAA), reused.currentQtypePrefer())
+	require.Equal(t, dnsmessage.TypeAAAA, reused.currentQtypePrefer())
 	enabled, ttl, maxCacheSize := reused.currentOptimisticCacheConfig()
 	require.True(t, enabled)
 	require.Equal(t, 45, ttl)
